POVRayDiagram階層
